vaal-ai-empire/requirements.txt
qwen-agent>=0.0.6
oss2>=2.18.4
mistralai>=1.0.0
cohere>=5.5.0
pandas>=2.2.0
python-dotenv>=1.0.1
dashscope>=1.20.0
huggingface_hub>=0.23.0
pypdf>=4.2.0
jira>=3.8.0
notion-client>=2.2.1
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
PyGithub>=2.3.0
yfinance>=0.2.40
ta>=0.11.0
python-wordpress-xmlrpc>=2.3
mailchimp-marketing>=3.0.80
